+++ Sources/TuistDependencies/DependenciesController.swift
@@ -167,9 +167,7 @@ public final class DependenciesController: DependenciesControlling {
             try carthageInteractor.clean(dependenciesDirectory: dependenciesDirectory)
         }
 
-        if let swiftPackageManagerDependencies = dependencies.swiftPackageManager,
-           !swiftPackageManagerDependencies.packages.isEmpty
-        {
+        if let swiftPackageManagerDependencies = dependencies.swiftPackageManager {
             let swiftPackageManagerDependenciesGraph = try swiftPackageManagerInteractor.install(
                 dependenciesDirectory: dependenciesDirectory,
                 dependencies: swiftPackageManagerDependencies,