+++ Sources/TuistGraph/Models/LaunchStyle.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum LaunchStyle: Codable {
+    case automatically
+    case waitForExecutableToBeLaunched
+}