+++ Sources/TuistGraph/Models/Metadata/XCFrameworkMetadata.swift
@@ -8,18 +8,21 @@ public struct XCFrameworkMetadata: Equatable {
     public var primaryBinaryPath: AbsolutePath
     public var linking: BinaryLinking
     public var mergeable: Bool
+    public var status: FrameworkStatus
 
     public init(
         path: AbsolutePath,
         infoPlist: XCFrameworkInfoPlist,
         primaryBinaryPath: AbsolutePath,
         linking: BinaryLinking,
-        mergeable: Bool
+        mergeable: Bool,
+        status: FrameworkStatus
     ) {
         self.path = path
         self.infoPlist = infoPlist
         self.primaryBinaryPath = primaryBinaryPath
         self.linking = linking
         self.mergeable = mergeable
+        self.status = status
     }
 }