+++ Sources/TuistGraphTesting/Models/ResourceSynthesizers+TestData.swift
@@ -4,9 +4,10 @@ import Foundation
 extension TuistGraph.ResourceSynthesizer {
     public static func test(
         parser: Parser = .assets,
+        parserOptions: [String: Parser.Option] = [:],
         extensions: Set<String> = ["xcassets"],
         template: Template = .defaultTemplate("Assets")
     ) -> Self {
-        ResourceSynthesizer(parser: parser, extensions: extensions, template: template)
+        ResourceSynthesizer(parser: parser, parserOptions: parserOptions, extensions: extensions, template: template)
     }
 }