+++ Tests/TuistCoreTests/Graph/GraphLoaderTests.swift
@@ -174,8 +174,8 @@ final class GraphLoaderTests: TuistUnitTestCase {
 
     func test_loadWorkspace_frameworkDependency() throws {
         // Given
-        let targetA = Target.test(name: "A", dependencies: [.framework(path: "/Frameworks/F1.framework")])
-        let targetB = Target.test(name: "B", dependencies: [.framework(path: "/Frameworks/F2.framework")])
+        let targetA = Target.test(name: "A", dependencies: [.framework(path: "/Frameworks/F1.framework", status: .required)])
+        let targetB = Target.test(name: "B", dependencies: [.framework(path: "/Frameworks/F2.framework", status: .required)])
         let projectA = Project.test(path: "/A", name: "A", targets: [targetA])
         let projectB = Project.test(path: "/B", name: "B", targets: [targetB])
         let workspace = Workspace.test(path: "/", name: "Workspace", projects: ["/A", "/B"])
@@ -216,7 +216,8 @@ final class GraphLoaderTests: TuistUnitTestCase {
                     bcsymbolmapPaths: [],
                     linking: .dynamic,
                     architectures: [.arm64],
-                    isCarthage: false
+                    isCarthage: false,
+                    status: .required
                 ),
             ]),
             .target(name: "B", path: "/B"): Set([
@@ -227,7 +228,8 @@ final class GraphLoaderTests: TuistUnitTestCase {
                     bcsymbolmapPaths: [],
                     linking: .static,
                     architectures: [.x8664],
-                    isCarthage: false
+                    isCarthage: false,
+                    status: .required
                 ),
             ]),
         ])
@@ -235,8 +237,8 @@ final class GraphLoaderTests: TuistUnitTestCase {
 
     func test_loadWorkspace_frameworkDependencyReferencedMultipleTimes() throws {
         // Given
-        let targetA = Target.test(name: "A", dependencies: [.framework(path: "/Frameworks/F.framework")])
-        let targetB = Target.test(name: "B", dependencies: [.framework(path: "/Frameworks/F.framework")])
+        let targetA = Target.test(name: "A", dependencies: [.framework(path: "/Frameworks/F.framework", status: .required)])
+        let targetB = Target.test(name: "B", dependencies: [.framework(path: "/Frameworks/F.framework", status: .required)])
         let projectA = Project.test(path: "/A", name: "A", targets: [targetA])
         let projectB = Project.test(path: "/B", name: "B", targets: [targetB])
         let workspace = Workspace.test(path: "/", name: "Workspace", projects: ["/A", "/B"])
@@ -268,7 +270,8 @@ final class GraphLoaderTests: TuistUnitTestCase {
             bcsymbolmapPaths: [],
             linking: .dynamic,
             architectures: [.arm64],
-            isCarthage: false
+            isCarthage: false,
+            status: .required
         )
         XCTAssertEqual(graph.dependencies, [
             .target(name: "A", path: "/A"): Set([
@@ -351,7 +354,10 @@ final class GraphLoaderTests: TuistUnitTestCase {
 
     func test_loadWorkspace_xcframeworkDependency() throws {
         // Given
-        let targetA = Target.test(name: "A", dependencies: [.xcframework(path: "/XCFrameworks/XF1.xcframework")])
+        let targetA = Target.test(
+            name: "A",
+            dependencies: [.xcframework(path: "/XCFrameworks/XF1.xcframework", status: .required)]
+        )
         let projectA = Project.test(path: "/A", name: "A", targets: [targetA])
         let workspace = Workspace.test(path: "/", name: "Workspace", projects: ["/A"])
 
@@ -361,7 +367,8 @@ final class GraphLoaderTests: TuistUnitTestCase {
                 infoPlist: .test(),
                 primaryBinaryPath: "/XCFrameworks/XF1.xcframework/ios-arm64/XF1",
                 linking: .dynamic,
-                mergeable: false
+                mergeable: false,
+                status: .required
             )
         )
 
@@ -383,7 +390,8 @@ final class GraphLoaderTests: TuistUnitTestCase {
                     infoPlist: .test(),
                     primaryBinaryPath: "/XCFrameworks/XF1.xcframework/ios-arm64/XF1",
                     linking: .dynamic,
-                    mergeable: false
+                    mergeable: false,
+                    status: .required
                 ),
             ]),
         ])
@@ -391,7 +399,10 @@ final class GraphLoaderTests: TuistUnitTestCase {
 
     func test_loadWorkspace_mergeableXCFrameworkDependency() throws {
         // Given
-        let targetA = Target.test(name: "A", dependencies: [.xcframework(path: "/XCFrameworks/XF1.xcframework")])
+        let targetA = Target.test(
+            name: "A",
+            dependencies: [.xcframework(path: "/XCFrameworks/XF1.xcframework", status: .required)]
+        )
         let projectA = Project.test(path: "/A", name: "A", targets: [targetA])
         let workspace = Workspace.test(path: "/", name: "Workspace", projects: ["/A"])
 
@@ -401,7 +412,8 @@ final class GraphLoaderTests: TuistUnitTestCase {
                 infoPlist: .test(),
                 primaryBinaryPath: "/XCFrameworks/XF1.xcframework/ios-arm64/XF1",
                 linking: .dynamic,
-                mergeable: true
+                mergeable: true,
+                status: .required
             )
         )
 
@@ -423,7 +435,8 @@ final class GraphLoaderTests: TuistUnitTestCase {
                     infoPlist: .test(),
                     primaryBinaryPath: "/XCFrameworks/XF1.xcframework/ios-arm64/XF1",
                     linking: .dynamic,
-                    mergeable: true
+                    mergeable: true,
+                    status: .required
                 ),
             ]),
         ])
@@ -675,7 +688,8 @@ final class GraphLoaderTests: TuistUnitTestCase {
                 bcsymbolmapPaths: [],
                 linking: metadata.linkage,
                 architectures: metadata.architectures,
-                isCarthage: false
+                isCarthage: false,
+                status: .required
             )
         }
         return provider