+++ Tests/TuistGeneratorTests/Generator/SchemeDescriptorsGeneratorTests.swift
@@ -536,6 +536,59 @@ final class SchemeDescriptorsGeneratorTests: XCTestCase {
         XCTAssertEqual(result.macroExpansion?.buildableIdentifier, "primary")
     }
 
+    func test_schemeLaunchAction_with_launchStyle() throws {
+        // Given
+        let projectPath = try AbsolutePath(validating: "/Project")
+        let app = Target.test(name: "App", product: .app)
+        let framework = Target.test(name: "Framework", product: .framework)
+
+        let runAction = RunAction.test(
+            executable: TargetReference(projectPath: projectPath, name: "App"),
+            filePath: projectPath,
+            expandVariableFromTarget: TargetReference(projectPath: projectPath, name: "Framework"),
+            launchStyle: .waitForExecutableToBeLaunched
+        )
+        let scheme = Scheme.test(name: "Scheme", runAction: runAction)
+        let project = Project.test(targets: [app, framework], schemes: [scheme])
+
+        let graph = Graph.test(
+            projects: [project.path: project],
+            targets: [
+                project.path: [
+                    app.name: app,
+                    framework.name: framework,
+                ],
+            ],
+            dependencies: [
+                .target(name: app.name, path: projectPath): [
+                    .target(name: framework.name, path: projectPath),
+                ],
+            ]
+        )
+        let graphTraverser = GraphTraverser(graph: graph)
+
+        // When
+        let got = try subject.schemeLaunchAction(
+            scheme: scheme,
+            graphTraverser: graphTraverser,
+            rootPath: projectPath,
+            generatedProjects: createGeneratedProjects(projects: [project])
+        )
+
+        // Then
+        let result = try XCTUnwrap(got)
+
+        XCTAssertEqual(result.buildConfiguration, "Debug")
+        XCTAssertEqual(result.selectedDebuggerIdentifier, "Xcode.DebuggerFoundation.Debugger.LLDB")
+        XCTAssertEqual(result.selectedLauncherIdentifier, "Xcode.DebuggerFoundation.Launcher.LLDB")
+        XCTAssertEqual(result.launchStyle, .wait)
+
+        XCTAssertEqual(result.macroExpansion?.buildableName, "Framework.framework")
+        XCTAssertEqual(result.macroExpansion?.blueprintName, "Framework")
+        XCTAssertEqual(result.macroExpansion?.referencedContainer, "container:Project.xcodeproj")
+        XCTAssertEqual(result.macroExpansion?.buildableIdentifier, "primary")
+    }
+
     func test_schemeTestAction_with_codeCoverageTargets() throws {
         // Given
         let projectPath = try AbsolutePath(validating: "/somepath/Project")