+++ Tests/TuistGeneratorTests/Utils/Mocks/MockSynthesizedResourceInterfaceGenerator.swift
@@ -3,14 +3,22 @@ import TuistGraph
 @testable import TuistGenerator
 
 final class MockSynthesizedResourceInterfaceGenerator: SynthesizedResourceInterfacesGenerating {
-    var renderStub: ((ResourceSynthesizer.Parser, String, String, String?, [AbsolutePath]) throws -> String)?
+    var renderStub: ((
+        ResourceSynthesizer.Parser,
+        [String: ResourceSynthesizer.Parser.Option],
+        String,
+        String,
+        String?,
+        [AbsolutePath]
+    ) throws -> String)?
     func render(
         parser: ResourceSynthesizer.Parser,
+        parserOptions: [String: ResourceSynthesizer.Parser.Option],
         templateString: String,
         name: String,
         bundleName: String?,
         paths: [AbsolutePath]
     ) throws -> String {
-        try renderStub?(parser, templateString, name, bundleName, paths) ?? ""
+        try renderStub?(parser, parserOptions, templateString, name, bundleName, paths) ?? ""
     }
 }