+++ Tests/TuistLoaderTests/Loaders/DependenciesModelLoaderTests.swift
@@ -33,7 +33,6 @@ final class DependenciesModelLoaderTests: TuistUnitTestCase {
     func test_loadDependencies() throws {
         // Given
         let temporaryPath = try temporaryPath()
-        let localSwiftPackagePath = temporaryPath.appending(component: "LocalPackage")
         let plugins = Plugins.test()
 
         manifestLoader.loadDependenciesStub = { _ in
@@ -42,12 +41,7 @@ final class DependenciesModelLoaderTests: TuistUnitTestCase {
                     .github(path: "Dependency1", requirement: .exact("1.1.1")),
                     .git(path: "Dependency1", requirement: .exact("2.3.4")),
                 ],
-                swiftPackageManager: .init(
-                    [
-                        .local(path: Path(localSwiftPackagePath.pathString)),
-                        .remote(url: "RemoteUrl.com", requirement: .exact("1.2.3")),
-                    ]
-                ),
+                swiftPackageManager: .init(),
                 platforms: [.iOS, .macOS]
             )
         }
@@ -64,10 +58,7 @@ final class DependenciesModelLoaderTests: TuistUnitTestCase {
                 ]
             ),
             swiftPackageManager: .init(
-                [
-                    .local(path: localSwiftPackagePath),
-                    .remote(url: "RemoteUrl.com", requirement: .exact("1.2.3")),
-                ],
+                .manifest,
                 productTypes: [:],
                 baseSettings: .init(configurations: [
                     .debug: .init(settings: [:], xcconfig: nil),